x/exp/types: extract Entity type with UnmarshalJSONWithSchema
Move entity parsing logic from the private unmarshalEntity function into
Entity.UnmarshalJSONWithSchema, making it available as a public API.
EntityMap.UnmarshalJSONWithSchema now delegates to Entity for each
element in the array.
